cc: Fix up lambda-cc and change command line usage
It now supports proper sub-CC detection:
lambda-cc <actual cc> [options]

It even works with CC="lambda-cc 'ccache clang'"

But the above case does not work to configure EFL. Autotools
fails somewhere, for some very strange reason (can not
recognise the compiler as being a GNU compiler).

Also, I am not a big fan of the piping as it rewrites
the whole GCC command line, modifying the arguments for
the source file and the output file. Also, we probably lose
the __FILE__ information, which is quite bad.

It's better to just use lambda-pp as a preprocessor only.